requirements.txt
opencv_python==4.12.0.88
Pillow==12.1.1
streamlit==1.52.2
ultralytics==8.4.15
